backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
python-multipart>=0.0.9
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
